Configure ruff + mypy + mutmut and pass the package clean
Add the static-analysis spine the repo lacked: a curated ruff lint set
(E,W,F,I,B,UP,S,SIM,RUF) + formatter, a strict-leaning mypy config, and a
mutmut config scoped to the deterministic engine. Declare python-dotenv
(imported in llm.py but previously only a transitive dep), consolidate the
dev tools into one dependency-group, and pin the interpreter (.python-version).

Then bring the tree to a clean pass: ruff autofixes (Optional -> X|None,
unused imports, isort, ClassVar for mutable class defaults, zip(strict=))
and real type annotations across agents/orchestrator/app/seed/persona/engine
(untyped params, Any-leaks at JSON boundaries, str-vs-Cluster mismatches).
Route handlers gain explicit return types, which double as FastAPI
response_model. No behaviour change — `uv run pytest` stays at 199 passed.

+++ tests/test_engine_props.py
@@ -17,7 +17,6 @@ from __future__ import annotations
 
 import math
 
-import numpy as np
 import pytest
 from hypothesis import HealthCheck, given, settings
 from hypothesis import strategies as st
@@ -221,7 +220,7 @@ def test_min_volume_empty_stages_is_unreachable_sentinel():
 )
 def test_posterior_mean_lies_between_prior_and_sample(pa, pb, n, frac):
     prior = (pa, pb)
-    k = int(round(frac * n))
+    k = round(frac * n)
     post = update_posterior(prior, k, n)
     prior_m = posterior_mean(prior)
     sample_m = k / n
@@ -284,8 +283,7 @@ def _candidate(gid, impact, ltv, days, p_now, p_plus_one, **kw):
 )
 def test_rank_queue_normalizes_marginal_gain_into_unit_interval(gains):
     cands = [
-        _candidate(f"g{i}", impact=0.5, ltv=0.5, days=30, p_now=pn, p_plus_one=pp)
-        for i, (pn, pp) in enumerate(gains)
+        _candidate(f"g{i}", impact=0.5, ltv=0.5, days=30, p_now=pn, p_plus_one=pp) for i, (pn, pp) in enumerate(gains)
     ]
     ranked = rank_queue(cands)
     norms = [r.marginal_p_gain for r in ranked]
